move initialization to main file
